Add docs examples for declare v2, v3 and deploy contract v3
