+++ starknet_py/tests/e2e/contract_interaction/deploy_test.py
@@ -6,7 +6,7 @@ import pytest
 
 from starknet_py.common import create_sierra_compiled_contract
 from starknet_py.contract import Contract, DeclareResult
-from starknet_py.net.client_models import InvokeTransactionV1, InvokeTransactionV3
+from starknet_py.net.client_models import InvokeTransactionV1
 from starknet_py.net.models import DeclareV2
 from starknet_py.tests.e2e.fixtures.constants import MAX_FEE, MAX_RESOURCE_BOUNDS_L1
 from starknet_py.tests.e2e.fixtures.misc import load_contract
@@ -116,35 +116,6 @@ async def test_deploy_contract_v1(account, cairo1_hello_starknet_class_hash: int
 
     transaction = await account.client.get_transaction(tx_hash=deploy_result.hash)
     assert isinstance(transaction, InvokeTransactionV1)
-
-    class_hash = await account.client.get_class_hash_at(
-        contract_address=contract.address
-    )
-    assert class_hash == cairo1_hello_starknet_class_hash
-
-
-@pytest.mark.asyncio
-async def test_deploy_contract_v3(account, cairo1_hello_starknet_class_hash: int):
-    compiled_contract = load_contract("HelloStarknet")["sierra"]
-    abi = create_sierra_compiled_contract(
-        compiled_contract=compiled_contract
-    ).parsed_abi
-
-    deploy_result = await Contract.deploy_contract_v3(
-        class_hash=cairo1_hello_starknet_class_hash,
-        account=account,
-        abi=abi,
-        l1_resource_bounds=MAX_RESOURCE_BOUNDS_L1,
-        cairo_version=1,
-    )
-    await deploy_result.wait_for_acceptance()
-
-    contract = deploy_result.deployed_contract
-    assert isinstance(contract.address, int)
-    assert len(contract.functions) != 0
-
-    transaction = await account.client.get_transaction(tx_hash=deploy_result.hash)
-    assert isinstance(transaction, InvokeTransactionV3)
 
     class_hash = await account.client.get_class_hash_at(
         contract_address=contract.address
